+++ src/components/checkbox/checkbox.tsx
@@ -225,7 +225,7 @@ export function Checkbox<
           form={() => props.form}
           name={() => props.name!}
           onReset={control.reset}
-          value={() => props.value || "on"}
+          value={() => props.value ?? "on"}
         />
       </Show>
       {renderElement({